Refactor project structure by separating logic, plotting, and user profile management into distinct modules
- Removed `logic.py`, `main.py`, and `plotting.py` files.
- Created dedicated modules for hydration calculations, performance modeling, weather data parsing, user profile management, and utility functions.
- Moved relevant functions to `logic/hydration.py`, `logic/performance.py`, `logic/weather_parser.py`, `logic/user_profiles.py`, and `logic/utils.py`.
- Created new plotting modules: `plotting/area_charts.py`, `plotting/line_charts.py`, `plotting/radar_charts.py`, and `plotting/gantt.py` for better organization.
- Added a new JSON file `data/user_profiles.json` to manage user profiles.
- Updated function signatures and docstrings to reflect new module organization.
