+++ README.md
@@ -38,12 +38,19 @@ MoveGuiderAI is built with a modular and scalable Python architecture designed f
 
 The project is structured with a clear separation of concerns, making it easy to maintain and debug:
 
-*   `app.py`: The main Streamlit application. It handles UI layout, state management (`st.session_state`), and orchestrates calls to the logic and plotting modules.
-*   `api_clients.py`: Contains all functions responsible for communicating with external APIs (Nominatim, WeatherAPI). This isolates external dependencies.
-*   `logic.py`: The "brains" of the application. This module contains all business logic, data processing, and modeling functions (e.g., parsing API responses, calculating hydration needs, modeling the energy curve).
-*   `plotting.py`: Contains all data visualization functions. Each function takes a clean Pandas DataFrame as input and returns a Plotly Figure object, completely decoupling charting from data logic.
-*   `config.py`: Loads and provides access to configuration and secret keys from the `.env` file.
-*   `.env`: A local, untracked file for storing secret API keys.
+* `app.py`: The main Streamlit application. It handles UI layout, state management (`st.session_state`), and orchestrates calls to the logic and plotting modules.
+* `api_clients.py`: Contains all functions responsible for communicating with external APIs (Nominatim, WeatherAPI).
+* `logic/`: A package containing all business logic, data processing, and modeling modules.
+  * `weather_parser.py`: Parses raw API data into clean DataFrames.
+  * `performance.py`: Models the user's energy curve.
+  * `hydration.py`: Calculates personalized hydration needs.
+  * `planner.py`: Handles daily routine logic, data unification, and recommendations.
+  * `user_profiles.py`: Manages loading and saving of user profiles and routines.
+  * `utils.py`: Contains common utility functions (e.g., timezone conversions).
+* `plotting/`: A package containing all data visualization functions. Each module returns a Plotly Figure object.
+* `data/`: Contains data files, such as the `user_profiles.json` for storing user settings.
+* `config.py`: Loads and provides access to configuration and secret keys from the `.env` file.
+* `.env`: A local, untracked file for storing secret API keys.
 
 ---
