+++ plotting/gantt.py
@@ -1,0 +1,5 @@
+# plotting/gantt.py
+
+# This file is a placeholder for the new Gantt chart plotting function.
+# We will implement this in the next step.
+pass
